feat: update create subtraction dialog styling
